LangchainAcademyLangGraph/pyproject.toml
[project]
name = "langchainacademylanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.6",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.1",
    "langgraph-prebuilt>=0.5.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
